CF update: keep `masked_sorted_indices` in `selection.util`.
